+++ config/webpack.config.js
@@ -21,7 +21,7 @@ const entryFiles = {
   '1995-edu-benefits': './src/js/edu-benefits/1995/edu-benefits-entry.jsx',
   '5490-edu-benefits': './src/js/edu-benefits/5490/edu-benefits-entry.jsx',
   '5495-edu-benefits': './src/js/edu-benefits/5495/edu-benefits-entry.jsx',
-  '526-claims-increase': './src/js/526/form-entry.jsx',
+  '526EZ-claims-increase': './src/js/526EZ/form-entry.jsx',
   'claims-status': './src/js/claims-status/claims-status-entry.jsx',
   'discharge-upgrade-instructions': './src/js/discharge-wizard/discharge-wizard-entry.jsx',
   'health-records': './src/js/health-records/health-records-entry.jsx',
