+++ src/js/disability-benefits/526EZ/components/IntroductionPage.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { focusElement } from '../../common/utils/helpers';
-import OMBInfo from '../../common/components/OMBInfo';
-import FormTitle from '../../common/schemaform/components/FormTitle';
+import { focusElement } from '../../../common/utils/helpers';
+import OMBInfo from '../../../common/components/OMBInfo';
+import FormTitle from '../../../common/schemaform/components/FormTitle';
 
-import SaveInProgressIntro, { introActions, introSelector } from '../../common/schemaform/save-in-progress/SaveInProgressIntro';
+import SaveInProgressIntro, { introActions, introSelector } from '../../../common/schemaform/save-in-progress/SaveInProgressIntro';
 
 class IntroductionPage extends React.Component {
   componentDidMount() {
