+++ src/js/disability-benefits/526EZ/helpers.js
@@ -1,4 +1,4 @@
-import { transformForSubmit } from '../common/schemaform/helpers';
+import { transformForSubmit } from '../../common/schemaform/helpers';
 
 export function transform(formConfig, form) {
   const formData = transformForSubmit(formConfig, form);
