+++ src/js/disability-benefits/526EZ/routes.jsx
@@ -1,4 +1,4 @@
-import { createRoutes as createFormRoutes } from '../common/schemaform/helpers';
+import { createRoutes as createFormRoutes } from '../../common/schemaform/helpers';
 import Form526EZApp from './Form526EZApp';
 import formConfig from './config/form';
